Agent: Extract method _move_file

+++ monkey/infection_monkey/dropper.py
@@ -77,27 +77,9 @@ class MonkeyDrops(object):
             os.remove(destination_path)
 
         # always try to move the file first
-        if not file_exists:
-            try:
-                shutil.move(self._config["source_path"], self._config["destination_path"])
+        if not file_exists and not self._move_file(source_path, destination_path):
+            # if file still need to change path, copy it
 
-                logger.info(
-                    "Moved source file '%s' into '%s'",
-                    self._config["source_path"],
-                    self._config["destination_path"],
-                )
-
-                file_exists = True
-            except (WindowsError, IOError, OSError) as exc:
-                logger.debug(
-                    "Error moving source file '%s' into '%s': %s",
-                    self._config["source_path"],
-                    self._config["destination_path"],
-                    exc,
-                )
-
-        # if file still need to change path, copy it
-        if not file_exists:
             try:
                 shutil.copy(self._config["source_path"], self._config["destination_path"])
 
@@ -180,6 +162,19 @@ class MonkeyDrops(object):
         time.sleep(3)
         if monkey_process.poll() is not None:
             logger.warning("Seems like monkey died too soon")
+
+    def _move_file(self, source_path, destination_path) -> bool:
+        try:
+            shutil.move(source_path, destination_path)
+            logger.info(f"Moved source file '{source_path}' into '{destination_path}'")
+        except (WindowsError, IOError, OSError) as exc:
+            logger.debug(
+                f"Error moving source file '{source_path}' into '{destination_path}': {exc}"
+            )
+
+            return False
+
+        return True
 
     def cleanup(self):
         logger.info("Cleaning up the dropper")
